fix(WinUI3): Apply PR suggestion
Only minor changes
