--- src/gui4/windows/kDrive client/kDrive client/Converters/ListEmptyToBooleanConverter.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Data;
-
-namespace Infomaniak.kDrive.Converters
-{
-    public class ListEmptyToBooleanConverter : IValueConverter
-    {
-       public object Convert(object value, Type targetType, object parameter, string language)
-        {
-            if (value is IEnumerable<object> list)
-            {
-                return !list.Any();
-            }
-            Logger.Log(Logger.Level.Fatal, "ListEmptyToBooleanConverter: value is not a list.");
-            throw new ArgumentException("Invalid value type", nameof(value));
-        }
-
-       public object ConvertBack(object value, Type targetType, object parameter, string language)
-        {
-            Logger.Log(Logger.Level.Fatal, "ListEmptyToBooleanConverter: ConvertBack is not supported.");
-            throw new NotSupportedException("ConvertBack is not supported.");
-        }
-    }
-}
